+++ src/cli/commands.rs
@@ -48,6 +48,7 @@ pub fn dispatch(cmd: Cmd) -> Result<()> {
             kill_timeout,
             max_log_size,
             no_log_file,
+            disable_logs,
             health_check,
             args,
         } => {
@@ -126,6 +127,7 @@ pub fn dispatch(cmd: Cmd) -> Result<()> {
                 app.max_log_size = Some(config::parse_memory(&s)?);
             }
             app.disable_log_files = no_log_file;
+            app.disable_logs = disable_logs;
             if let Some(cmd) = health_check {
                 app = app.health_check(cmd);
             }